[Edited]:เปลี่ยน type ของ price จาก int -> float แทน

+++ book.go
@@ -3,11 +3,11 @@ package book
 type Book struct {
 	ID    string
 	Name  string
-	Price int
+	Price float64
 	Stock int
 }
 
-func NewBook(name string, price int) Book {
+func NewBook(name string, price float64) Book {
 	return Book{
 		Name:  name,
 		Price: price,
@@ -18,6 +18,6 @@ func (b Book) getName() string {
 	return b.Name
 }
 
-func (b Book) getPrice() int {
+func (b Book) getPrice() float64 {
 	return b.Price
 }
